Ensure the collection script to find test results always runs (#36)
The issue when not specifying `when: always` for that command is that although store_test_results and store_artifacts implicitly always runs even if the build is failing, homegrown commands like the one provided here don't. Thus the need to add `when: always` to force it to run regardless of build status.